refactor: :recycle: extract styled table cell to a reusable component
